backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.38.0
openai==1.40.0
azure-ai-contentsafety==1.0.0
azure-core==1.30.2
